turn zoom into extra component and util
